refactor: apply service abstraction for user layer
